signer: Refactor User into a file of its own
