+++ playground/signer/playground_sign/_signer_repository.py
@@ -2,9 +2,7 @@
 
 """Internal repository module for playground signer tool"""
 
-from configparser import ConfigParser
 from contextlib import AbstractContextManager
-import sys
 import click
 import filecmp
 import json
@@ -38,6 +36,7 @@ from tuf.api.metadata import (
 from tuf.api.serialization.json import CanonicalJSONSerializer, JSONSerializer
 from tuf.repository import Repository, AbortEdit
 
+from playground_sign._user import User
 
 logger = logging.getLogger(__name__)
 
@@ -72,95 +71,8 @@ class OfflineConfig:
     signing_period: int
 
 
-class User:
-    """Class that manages user configuration and stores the signer cache for the user"""
-
-    def __init__(self, path: str):
-        self._config_path = path
-
-        self._config = ConfigParser(interpolation=None)
-        self._config.read(path)
-
-        # TODO: create config if missing, ask/confirm values from user
-        if not self._config:
-            raise click.ClickException(f"Settings file {path} not found")
-        try:
-            self.name = self._config["settings"]["user-name"]
-            self.pykcs11lib = self._config["settings"]["pykcs11lib"]
-            self.push_remote = self._config["settings"]["push-remote"]
-            self.pull_remote = self._config["settings"]["pull-remote"]
-        except KeyError as e:
-            raise click.ClickException(f"Failed to find required setting {e} in {path}")
-
-        # signing key config is not required
-        if "signing-keys" in self._config:
-            self._signing_key_uris = dict(self._config.items("signing-keys"))
-        else:
-            self._signing_key_uris = {}
-
-        # signer cache gets populated as they are used the first time
-        self._signers: dict[str, Signer] = {}
-
-    def get_signer(self, key: Key) -> Signer:
-        """Returns a Signer for the given public key
-
-        The signer sources are (in order):
-        * any configured signer from 'signing-keys' config section
-        * for sigstore type keys, a Signer is automatically created
-        * for any remaining keys, HSM is assumed and a signer is created
-        """
-
-        def get_secret(secret: str) -> str:
-            msg = f"Enter {secret} to sign"
-
-            # special case for tests -- prompt() will lockup trying to hide STDIN:
-            if not sys.stdin.isatty():
-                return sys.stdin.readline().rstrip()
-
-            return click.prompt(bold(msg), hide_input=True)
-
-        if key.keyid in self._signers:
-            # signer is already cached
-            pass
-        elif key.keyid in self._signing_key_uris:
-            # signer is not cached yet, but the uri was configured
-            uri = self._signing_key_uris[key.keyid]
-            self._signers[key.keyid] = Signer.from_priv_key_uri(uri, key, get_secret)
-
-        elif key.keytype == "sigstore-oidc":
-            # signer is not cached, no configuration was found, type is sigstore
-            self._signers[key.keyid] = Signer.from_priv_key_uri(
-                "sigstore:?ambient=false", key, get_secret
-            )
-        else:
-            # signer is not cached, no configuration was found: assume Yubikey
-            self._signers[key.keyid] = Signer.from_priv_key_uri("hsm:", key, get_secret)
-
-        return self._signers[key.keyid]
-
-    def store_signer(self, uri: str, key: Key) -> None:
-        """Store the uri in the signing-keys section of the config file"""
-
-        # NOTE Currently we set the uri and commit to a file immediately:
-        # it might make sense to separate these steps so that we could only commit
-        # after the whole process has finished (e.g. no git issues or something)
-
-        self._signing_key_uris[key.keyid] = uri
-
-        if "signing-keys" not in self._config:
-            self._config.add_section("signing-keys")
-        self._config["signing-keys"][key.keyid] = uri
-
-        with open(self._config_path, "w") as f:
-            self._config.write(f)
-
-
 def blue(text: str) -> str:
     return click.style(text, fg="bright_blue")
-
-
-def bold(text: str) -> str:
-    return click.style(text, bold=True)
 
 
 def _find_changed_roles(known_good_dir: str, signing_event_dir: str) -> list[str]:
@@ -320,7 +232,7 @@ class SignerRepository(Repository):
             for keyid in r.keyids:
                 keys.append(delegator.get_key(keyid))
         except ValueError:
-            pass # either no role delegated yet or key not found
+            pass  # either no role delegated yet or key not found
 
         return keys
